third_party/WebKit/LayoutTests/external/wpt/cssom/CSS-expected.txt

This is a testharness.js-based test.
PASS CSS.escape 
FAIL CSS.supports, one argument form assert_equals: CSS.supports: Single-argument form allows for declarations without enclosing parentheses expected true but got false
PASS CSS.supports, two argument form 
Harness: the test ran to completion.
